+++ libs/shared/host-health/package.json
@@ -1,0 +1,11 @@
+{
+    "name": "@iptvnator/shared/host-health",
+    "version": "0.0.1",
+    "private": true,
+    "type": "commonjs",
+    "main": "./src/index.js",
+    "types": "./src/index.d.ts",
+    "dependencies": {
+        "tslib": "^2.3.0"
+    }
+}
